refactor: move rules to another file
