+++ src/types.ts
@@ -1,0 +1,6 @@
+/**
+ * Options accepted by the presetForms preset
+ */
+export interface PresetFormOptions {
+  strategy?: 'base' | 'class'
+}
